+++ CHANGELOG.md
@@ -191,6 +191,10 @@ Volume manager:
 - For new node filters during volume creation, the merge mode can now be
   selected, i.e. whether additional filters will be combined using OR or AND.
 
+- The new "Import from CATMAID" tab allows users with import permission to
+  search and import volumes/meshes from other CATMAID projects, both local and
+  remote. Volumes can be searched by name and annotation.
+
 Graph widget:
 
 - Add options to configure the edge text outline size and opacity. The options